third_party/WebKit/LayoutTests/fast/forms/selection-wrongtype-expected.txt

Tests to ensure selectionStart, selectionEnd, selectionDirection and selectionRange throw exceptions when the input element is not a text field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('button') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('button') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('button') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('button') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('button') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('button') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('checkbox') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('checkbox') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('checkbox') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('checkbox') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('checkbox') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('checkbox') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('image') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('image') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('image') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('image') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('image') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('image') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('radio') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('radio') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('radio') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('radio') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('radio') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('radio') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('range') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('range') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('range') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('range') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('range') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('range') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('reset') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('reset') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('reset') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('reset') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('reset') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('reset') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('submit') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('submit') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('submit') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('submit') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('submit') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('submit') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('email') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('email') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('email') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('email') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('email') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('email') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('number') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('number') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('number') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('number') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('number') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('number') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('date') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('date') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('date') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('date') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('date') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('date') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('time') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('time') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('time') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('time') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('time') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('time') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('month') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('month') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('month') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('month') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('month') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('month') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('week') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('week') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('week') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('week') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('week') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('week') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('datetime-local') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('datetime-local') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('datetime-local') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('datetime-local') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('datetime-local') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('datetime-local') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS test.selectionStart threw exception InvalidStateError: Failed to read the 'selectionStart' property from 'HTMLInputElement': The input element's type ('color') does not support selection..
PASS test.selectionStart = 0 threw exception InvalidStateError: Failed to set the 'selectionStart' property on 'HTMLInputElement': The input element's type ('color') does not support selection..
PASS test.selectionEnd threw exception InvalidStateError: Failed to read the 'selectionEnd' property from 'HTMLInputElement': The input element's type ('color') does not support selection..
PASS test.selectionEnd = 0 threw exception InvalidStateError: Failed to set the 'selectionEnd' property on 'HTMLInputElement': The input element's type ('color') does not support selection..
PASS test.selectionDirection threw exception InvalidStateError: Failed to read the 'selectionDirection' property from 'HTMLInputElement': The input element's type ('color') does not support selection..
PASS test.selectionDirection = 0 threw exception InvalidStateError: Failed to set the 'selectionDirection' property on 'HTMLInputElement': The input element's type ('color') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function.
PASS successfullyParsed is true

TEST COMPLETE
